webpki: lift VerifierBuilderError to mod.rs
This commit moves the `VerifierBuilderError` from `client_verifier.rs`
to `mod.rs` where it can be shared between both the client and server
verifiers in the future.
